Bug: ERC-1155 Token Amount Validation Flaw
ERC-1155 amount validation incorrectly allows decimal inputs. The validation logic uses parseInt(value), which truncates decimals, then checks if the result is an integer. Since parseInt() always returns an integer (or NaN), this check always passes for valid numbers, failing to enforce the whole number requirement for ERC-1155 token amounts.

+++ source/pages/Send/SendEth.tsx
@@ -984,9 +984,9 @@ export const SendEth = () => {
 
                         // For ERC-1155 NFTs, validate against verified balance
                         if (isERC1155) {
-                          const numValue = parseInt(value);
+                          const numValue = parseFloat(value);
 
-                          // Must be a positive integer
+                          // Must be a positive integer (check the original float value)
                           if (!Number.isInteger(numValue) || numValue <= 0) {
                             return Promise.reject(
                               t('send.amountMustBePositiveInteger')
